mas refactor y añadido login y register
