+++ src/controllers/user_controller.js
@@ -1,0 +1,75 @@
+import db from "#src/models/index";
+import bcrypt from "bcryptjs"; // Importamos bcrypt para el hashing de contraseñas
+
+/**
+ * Registro de usuario
+ *
+ * - Recibe: `nickname`, `password`, `mail`, `style_fav`
+ * - Verifica si el correo ya está registrado.
+ * - Hashea la contraseña antes de almacenarla en la base de datos.
+ * - Devuelve un mensaje de éxito con los datos del usuario registrado.
+ */
+export const registerUser = async (req, res) => {
+    const { nickname, password, mail, style_fav } = req.body;
+
+    try {
+        // Verificar si el usuario ya existe en la base de datos
+        const userExists = await db.user.findOne({ where: { mail } });
+
+        console.log("Usuario encontrado:", userExists);
+
+        if (userExists) {
+            return res.status(400).json({ error: "Correo ya registrado" });
+        }
+
+        // Hashear la contraseña antes de guardarla
+        const salt = await bcrypt.genSalt(10); // Generamos un salt (valor aleatorio para mayor seguridad)
+        const hashedPassword = await bcrypt.hash(password, salt); // Hasheamos la contraseña con el salt
+
+        // Crear nuevo usuario con la contraseña encriptada
+        const newUser = await db.user.create({
+            nickname,
+            password: hashedPassword, // Guardamos la contraseña hasheada
+            mail,
+            style_fav
+        });
+
+        res.status(201).json({ message: "Usuario registrado con éxito", user: newUser });
+    } catch (error) {
+        console.error("Error al registrar usuario:", error);
+        res.status(500).json({ error: "Error al registrar usuario" });
+    }
+};
+
+/**
+ * Inicio de sesión de usuario
+ *
+ * - Recibe: `mail`, `password`
+ * - Busca el usuario en la base de datos por su correo.
+ * - Verifica la contraseña utilizando `bcrypt.compare()`.
+ * - Si es correcta, devuelve un mensaje de éxito.
+ * - Si es incorrecta, devuelve un error de autenticación.
+ */
+export const loginUser = async (req, res) => {
+    const { mail, password } = req.body;
+
+    try {
+        // Buscar usuario en la base de datos por su correo
+        const foundUser = await db.user.findOne({ where: { mail } });
+
+        if (!foundUser) {
+            return res.status(404).json({ error: "Usuario no encontrado" });
+        }
+
+        // Comparar la contraseña ingresada con la contraseña hasheada en la BD
+        const validPassword = await bcrypt.compare(password, foundUser.password);
+        if (!validPassword) {
+            return res.status(401).json({ error: "Contraseña incorrecta" });
+        }
+
+        res.status(200).json({ message: "Login exitoso", user: foundUser });
+    } catch (error) {
+        console.error("Error en el login:", error);
+        res.status(500).json({ error: "Error en el login" });
+    }
+};
